pyproject.toml
[tool.poetry]
name = "tutorialgen"
version = "0.1.0"
description = ""
authors = ["Toshimichi0915 <main@toshimichi.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
